+++ nextstep-di/src/main/java/nextstep/di/factory/BeanScanner.java
@@ -1,12 +1,12 @@
-package nextstep.mvc.tobe;
+package nextstep.di.factory;
 
-import nextstep.di.factory.BeanFactory;
 import org.reflections.Reflections;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 import java.lang.annotation.Annotation;
 import java.util.Arrays;
+import java.util.Collection;
 import java.util.Set;
 import java.util.stream.Collectors;
 
@@ -29,7 +29,7 @@ public class BeanScanner {
     private Set<Class<?>> scanBeans() {
         return Arrays.stream(types)
                 .map(type -> reflections.getTypesAnnotatedWith(type))
-                .flatMap(sets -> sets.stream())
+                .flatMap(Collection::stream)
                 .collect(Collectors.toSet());
     }
 }
